Extract things into top-level directory
Move a few modules from `util` to (for now) the top-level
directory. These modules have multiple dependencies, or are
otherwise application-specific, but don't fit comfortably in
the `helpers` modules.
