+++ src/sidebar/components/annotation-body.js
@@ -4,7 +4,7 @@ import propTypes from 'prop-types';
 
 import { useStoreProxy } from '../store/use-store';
 import { isHidden } from '../helpers/annotation-metadata';
-import { withServices } from '../util/service-context';
+import { withServices } from '../service-context';
 import { applyTheme } from '../helpers/theme';
 
 import Button from './button';
