+++ src/sidebar/components/annotation-editor.js
@@ -3,7 +3,7 @@ import { useState } from 'preact/hooks';
 import propTypes from 'prop-types';
 
 import { normalizeKeyName } from '../../shared/browser-compatibility-utils';
-import { withServices } from '../util/service-context';
+import { withServices } from '../service-context';
 import { applyTheme } from '../helpers/theme';
 import { useStoreProxy } from '../store/use-store';
