+++ src/sidebar/components/annotation-publish-control.js
@@ -4,7 +4,7 @@ import propTypes from 'prop-types';
 import { useStoreProxy } from '../store/use-store';
 import { isNew, isReply } from '../helpers/annotation-metadata';
 import { isShared } from '../helpers/permissions';
-import { withServices } from '../util/service-context';
+import { withServices } from '../service-context';
 import { applyTheme } from '../helpers/theme';
 
 import Button from './button';
