+++ src/sidebar/components/annotation-view.js
@@ -3,7 +3,7 @@ import { useEffect, useState } from 'preact/hooks';
 import propTypes from 'prop-types';
 
 import { useStoreProxy } from '../store/use-store';
-import { withServices } from '../util/service-context';
+import { withServices } from '../service-context';
 import useRootThread from './hooks/use-root-thread';
 
 import ThreadList from './thread-list';
