+++ src/sidebar/components/moderation-banner.js
@@ -4,7 +4,7 @@ import propTypes from 'prop-types';
 
 import { useStoreProxy } from '../store/use-store';
 import * as annotationMetadata from '../helpers/annotation-metadata';
-import { withServices } from '../util/service-context';
+import { withServices } from '../service-context';
 
 /**
  * @typedef {import('../../types/api').Annotation} Annotation
