+++ src/sidebar/components/new-note-btn.js
@@ -3,7 +3,7 @@ import propTypes from 'prop-types';
 
 import uiConstants from '../ui-constants';
 import { useStoreProxy } from '../store/use-store';
-import { withServices } from '../util/service-context';
+import { withServices } from '../service-context';
 import { applyTheme } from '../helpers/theme';
 
 import Button from './button';
