+++ src/sidebar/components/selection-tabs.js
@@ -4,7 +4,7 @@ import propTypes from 'prop-types';
 
 import { useStoreProxy } from '../store/use-store';
 import uiConstants from '../ui-constants';
-import { withServices } from '../util/service-context';
+import { withServices } from '../service-context';
 
 import NewNoteBtn from './new-note-btn';
 import SvgIcon from '../../shared/components/svg-icon';
