+++ src/sidebar/components/share-links.js
@@ -1,7 +1,7 @@
 import { createElement } from 'preact';
 import propTypes from 'prop-types';
 
-import { withServices } from '../util/service-context';
+import { withServices } from '../service-context';
 
 import SvgIcon from '../../shared/components/svg-icon';
