+++ src/sidebar/components/sidebar-view.js
@@ -3,7 +3,7 @@ import propTypes from 'prop-types';
 import { useEffect, useRef } from 'preact/hooks';
 
 import useRootThread from './hooks/use-root-thread';
-import { withServices } from '../util/service-context';
+import { withServices } from '../service-context';
 import { useStoreProxy } from '../store/use-store';
 import { tabForAnnotation } from '../helpers/tabs';
