+++ src/sidebar/components/stream-view.js
@@ -3,7 +3,7 @@ import { useCallback, useEffect } from 'preact/hooks';
 import propTypes from 'prop-types';
 
 import * as searchFilter from '../util/search-filter';
-import { withServices } from '../util/service-context';
+import { withServices } from '../service-context';
 import useRootThread from './hooks/use-root-thread';
 import { useStoreProxy } from '../store/use-store';
