+++ src/sidebar/components/tag-editor.js
@@ -2,7 +2,7 @@ import { createElement } from 'preact';
 import { useRef, useState } from 'preact/hooks';
 import propTypes from 'prop-types';
 
-import { withServices } from '../util/service-context';
+import { withServices } from '../service-context';
 
 import AutocompleteList from './autocomplete-list';
 import { normalizeKeyName } from '../../shared/browser-compatibility-utils';
