+++ src/sidebar/components/tag-list.js
@@ -3,7 +3,7 @@ import { useMemo } from 'preact/hooks';
 import propTypes from 'prop-types';
 
 import { isThirdPartyUser } from '../helpers/account-id';
-import { withServices } from '../util/service-context';
+import { withServices } from '../service-context';
 
 /** @typedef {import('../../types/api').Annotation} Annotation */
