+++ src/sidebar/components/thread-card.js
@@ -5,7 +5,7 @@ import { useCallback, useMemo } from 'preact/hooks';
 
 import propTypes from 'prop-types';
 import { useStoreProxy } from '../store/use-store';
-import { withServices } from '../util/service-context';
+import { withServices } from '../service-context';
 
 import Thread from './thread';
