+++ src/sidebar/components/toast-messages.js
@@ -3,7 +3,7 @@ import { createElement } from 'preact';
 import propTypes from 'prop-types';
 
 import { useStoreProxy } from '../store/use-store';
-import { withServices } from '../util/service-context';
+import { withServices } from '../service-context';
 
 import SvgIcon from '../../shared/components/svg-icon';
