+++ src/sidebar/services/test/oauth-auth-test.js
@@ -1,5 +1,5 @@
 import { Injector } from '../../../shared/injector';
-import FakeWindow from '../../util/test/fake-window';
+import FakeWindow from '../../test/fake-window';
 import authFactory, { $imports } from '../oauth-auth';
 
 const DEFAULT_TOKEN_EXPIRES_IN_SECS = 1000;
