+++ src/sidebar/test/fake-window.js
